chore: remove packages used for debugging when running tests
